Bugs found in code1.c and code2.c corrected
Two bugs were found in code1.c and code2.c. 
The particular cases are:
When the minimum number of time slots is 1 (the program may report 2)
Exceptional cases in which the minimum number of time slots is plus one or minus one that that of reported by the previous version. You can only identify these cases by repeating the experiment with the new codes using the same random seed. 
The running time of 1) should not be changed. In case 2), there might exist some small variation on the running time, since it may give rise to plus one or minus one recursive call. This variation should be, anyway, very small.